Refactor UI code
