+++ ui/remix.config.js
@@ -1,10 +1,7 @@
 /** @type {import('@remix-run/dev').AppConfig} */
 export default {
-  ignoredRouteFiles: ["**/.*"],
-  postcss: true,
-  // appDirectory: "app",
-  // assetsBuildDirectory: "public/build",
-  // publicPath: "/build/",
-  // serverBuildPath: "build/index.js",
-  serverDependenciesToBundle: ['react-cytoscapejs'],
+    appDirectory: 'src',
+    ignoredRouteFiles: ['**/.*'],
+    postcss: true,
+    serverDependenciesToBundle: ['react-cytoscapejs'],
 };
