+++ ui/src/components/style/Footer.module.css
@@ -1,22 +1,22 @@
 .footer {
     margin-top: rem(120px);
     border-top: rem(1px) solid light-dark(var(--mantine-color-gray-2), var(--mantine-color-dark-5));
-  }
-  
-  .inner {
+}
+
+.inner {
     display: flex;
     justify-content: space-between;
     align-items: center;
     padding: var(--mantine-spacing-md) var(--mantine-spacing-md);
-  
+
     @media (max-width: rem(48em)) {
-      flex-direction: column;
+    flex-direction: column;
     }
-  }
-  
-  .links {
+}
+
+.links {
     @media (max-width: rem(48em)) {
-      margin-top: var(--mantine-spacing-lg);
-      margin-bottom: var(--mantine-spacing-sm);
+    margin-top: var(--mantine-spacing-lg);
+    margin-bottom: var(--mantine-spacing-sm);
     }
-  }
+}
