+++ ui/src/components/style/Navbar.module.css
@@ -3,21 +3,21 @@
     margin-bottom: rem(120px);
     background-color: var(--mantine-color-body);
     border-bottom: rem(1px) solid light-dark(var(--mantine-color-gray-3), var(--mantine-color-dark-4));
-  }
-  
-  .inner {
+}
+
+.inner {
     height: rem(56px);
     display: flex;
     justify-content: space-between;
     align-items: left;
-  }
-  
-  .icon {
+}
+
+.icon {
     width: rem(22px);
     height: rem(22px);
-  }
+}
 
-  .dark {
+.dark {
     @mixin dark {
         display: none
     }
@@ -25,9 +25,9 @@
     @mixin light {
         display:block
     }
-  }
+}
 
-  .light {
+.light {
     @mixin light {
         display: none
     }
@@ -35,4 +35,4 @@
     @mixin dark {
         display: block;
     }
-  }
+}
