+++ ui/src/entry.client.tsx
@@ -8,10 +8,10 @@ import { startTransition, StrictMode } from "react";
 import { hydrateRoot } from "react-dom/client";
 
 startTransition(() => {
-  hydrateRoot(
-    document,
-    <StrictMode>
-      <RemixBrowser />
-    </StrictMode>
-  );
+    hydrateRoot(
+        document,
+        <StrictMode>
+            <RemixBrowser />
+        </StrictMode>
+    );
 });
